# otelslog: Split code attributes